+++ src/library/math/operator/predicate/exports.ts
